Refactor About and Display Properties to src/shell
- Move `src/apps/about` to `src/shell/about`
- Move `src/apps/display-properties` to `src/shell/display-properties`
- Create `index.js` for both with dedicated `launch` functions
- Update all call sites to use direct launch functions instead of `launchApp`
- Update `DesktopThemesApp` to use the new path for `theme-preview.js`
- Unregister them from the global application registry to treat them as internal shell components
