+++ src/shell/about/about-app.js
