+++ src/shell/about/index.js
@@ -1,0 +1,6 @@
+import { AboutApp } from './about-app.js';
+
+export async function launchAboutApp(data = null) {
+  const app = new AboutApp(AboutApp.config);
+  return app.launch(data);
+}
