+++ src/shell/display-properties/background/background.html
