+++ src/shell/display-properties/background/background.js
