+++ src/shell/display-properties/color-spectrum.css
