+++ src/shell/display-properties/display-properties-app.js
