+++ src/shell/display-properties/displayproperties.css
