+++ src/shell/display-properties/index.js
@@ -1,0 +1,6 @@
+import DisplayPropertiesApp from './display-properties-app.js';
+
+export async function launchDisplayPropertiesApp(data = null) {
+  const app = new DisplayPropertiesApp(DisplayPropertiesApp.config);
+  return app.launch(data);
+}
