+++ src/shell/display-properties/settings/settings.js
