+++ src/shell/display-properties/theme-preview.js
